apps/summarizer/requirements/base.txt
channels-redis
Celery
channels
Django==5.0.1
django-allauth==0.60.0

git+https://github.com/celery/django-celery-beat#egg=django-celery-beat
django-channels-graphql-ws
django-cors-headers
django-filter
django-graphql-jwt
django-redis
graphene
graphene-django
graphql-ws
paypal-checkout-serversdk
pdfkit
pytz
Pillow
psycopg2
redis
stripe